+++ philo/src/parse.c
@@ -6,13 +6,13 @@
 /*   By: pmolnar <pmolnar@student.codam.nl>           +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2022/06/07 12:54:06 by pmolnar       #+#    #+#                 */
-/*   Updated: 2022/07/25 14:13:56 by pmolnar       ########   odam.nl         */
+/*   Updated: 2022/07/26 23:04:36 by pmolnar       ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <philo.h>
 
-int	validate_arg(int argc, t_sim *data)
+uint16_t	validate_arg(int argc, t_sim *data)
 {
 	uint16_t	i;
 
@@ -28,7 +28,7 @@ int	validate_arg(int argc, t_sim *data)
 	return (EXIT_SUCCESS);
 }
 
-int	parse_args(int argc, char *argv[], t_sim *data)
+uint16_t	parse_args(int argc, char *argv[], t_sim *data)
 {	
 	uint16_t	i;
